UI-Rendering: next steps

+++ src/Engine/Assets/AssetRegistry.cpp
@@ -57,7 +57,7 @@ bool AssetRegistry::registerAssetPath( const String &mount, const String &path )
         return false;
     }
     const ui32 hashId( StringUtils::hashName( mount ) );
-    s_instance->m_name2pathMap.insert( (size_t) hashId, path );
+    s_instance->m_name2pathMap.insert( hashId, path );
 
     return true;
 }
